[REFACTOR] Push Windows Docker Engine check earlier
When exporting a container on Windows, the Docker daemon must be in
Windows mode. Previously, this check was buried deep in the logic,
after we had already done a lot of work to pull packages down. Now, we
perform this check first-thing, thereby short circuiting a whole bunch
of unnecessary work, and cleaning up a bunch of logic.

Signed-off-by: Christopher Maier <cmaier@chef.io>